+++ lua/neotest-golang/features/testify/init.lua
@@ -1,0 +1,8 @@
+local M = {}
+
+M.lookup = require("neotest-golang.features.testify.lookup")
+M.query = require("neotest-golang.features.testify.query")
+M.tree_modification =
+  require("neotest-golang.features.testify.tree_modification")
+
+return M
