+++ lua/neotest-golang/find.lua
@@ -6,9 +6,8 @@ local M = {}
 --- @param filename string
 --- @param start_path string
 --- @return string | nil
-function M.find_file_upwards(filename, start_path)
+function M.file_upwards(filename, start_path)
   local scan = require("plenary.scandir")
-  local cwd = vim.fn.getcwd()
   local found_filepath = nil
   while start_path ~= vim.fn.expand("$HOME") do
     local files = scan.scan_dir(
@@ -34,6 +33,21 @@ function M.find_file_upwards(filename, start_path)
   end
 
   return found_filepath
+end
+
+-- Get all *_test.go files in a directory recursively.
+-- FIXME: do not use `find`, as this puts unnecessary dependency here?
+-- Might want to use plenary instead.
+function M.go_test_filepaths(folderpath)
+  local files = {}
+  local function scan_dir(dir)
+    local p = io.popen('find "' .. dir .. '" -type f -name "*_test.go"')
+    for file in p:lines() do
+      table.insert(files, file)
+    end
+  end
+  scan_dir(folderpath)
+  return files
 end
 
 return M
